Move PipelineAdvisor to its own file
