+++ fedot/core/pipelines/pipeline_node_factory.py
@@ -1,7 +1,7 @@
 from random import choice
 from typing import Optional
 
-from fedot.core.optimisers.advisor import PipelineChangeAdvisor
+from fedot.core.pipelines.pipeline_advisor import PipelineChangeAdvisor
 from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.graph import OptNode
 from fedot.core.optimisers.opt_node_factory import OptNodeFactory
